requirements.txt
streamlit
langchain
langchain-community
pandas
transformers
torch
google-generativeai
groq
openai
anthropic
openpyxl
python-dotenv
nltk